Complete the essentiel for application

+++ src/main/java/com/bookshop/dto/BookDTO.java
@@ -1,12 +1,12 @@
 package com.bookshop.dto;
 
 public record BookDTO(
-    Long id,
-    String title,
-    String author,
-    Double price,
-    Integer stock,
-    String description,
-    Long categoryId,
-    String categoryName
+        Long id,
+        String title,
+        String author,
+        Float price,
+        Integer stock,
+        String description,
+        Long categoryId,
+        String categoryName
 ) {}
